+++ bindings/nodejs/lib/types/utils/numeric.ts
@@ -1,0 +1,7 @@
+// Copyright 2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Used for serialised numeric strings
+ */
+export declare type NumericString = string;