+++ bindings/nodejs/lib/types/wallet/account.ts
@@ -7,6 +7,7 @@ import type { OutputData } from './output';
 import type { Transaction } from './transaction';
 import { CoinType } from '../../client';
 import { HexEncodedString } from '../utils';
+import { Bech32Address } from '../block/address';
 
 /**
  * Account identifier
@@ -75,7 +76,7 @@ export interface SyncOptions {
      * Specific Bech32 encoded addresses of the account to sync, if addresses are provided,
      * then `address_start_index` will be ignored
      */
-    addresses?: string[];
+    addresses?: Bech32Address[];
     /**
      * Address index from which to start syncing addresses. 0 by default, using a higher index will be faster because
      * addresses with a lower index will be skipped, but could result in a wrong balance for that reason
@@ -198,7 +199,7 @@ export interface FilterOptions {
     /** Filter all outputs where the booked milestone index is above the specified timestamp */
     upperBoundBookedTimestamp?: number;
     /** Filter all outputs for the provided types (Basic = 3, Alias = 4, Foundry = 5, NFT = 6) */
-    outputTypes?: Uint8Array;
+    outputTypes?: number[];
     /** Return all alias outputs matching these IDs. */
     aliasIds?: AliasId[];
     /** Return all foundry outputs matching these IDs. */