+++ bindings/python/iota_sdk/client/_node_core_api.py
@@ -8,7 +8,7 @@ from iota_sdk.types.output import OutputWithMetadata, OutputMetadata
 from iota_sdk.types.output_id import OutputId
 from iota_sdk.types.payload import MilestonePayload
 from iota_sdk.types.utxo_changes import UtxoChanges
-from typing import List
+from typing import List, Union
 from dacite import from_dict
 
 
@@ -97,7 +97,8 @@ class NodeCoreAPI():
             'blockBytes': block_bytes
         })
 
-    def get_output(self, output_id: OutputId | HexStr) -> OutputWithMetadata:
+    def get_output(
+            self, output_id: Union[OutputId, HexStr]) -> OutputWithMetadata:
         """Get the output corresponding to the given output id.
 
         Returns:
@@ -109,8 +110,8 @@ class NodeCoreAPI():
             'outputId': output_id_str
         }))
 
-    def get_output_metadata(self, output_id: OutputId |
-                            HexStr) -> OutputMetadata:
+    def get_output_metadata(
+            self, output_id: Union[OutputId, HexStr]) -> OutputMetadata:
         """Get the output metadata corresponding to the given output id.
 
         Returns: