+++ cli/src/wallet.rs
@@ -9,7 +9,8 @@ use crate::{
     command::wallet::{
         accounts_command, add_account, backup_command, change_password_command, init_command,
         migrate_stronghold_snapshot_v2_to_v3_command, mnemonic_command, new_account_command, node_info_command,
-        restore_command, set_node_url_command, sync_command, unlock_wallet, InitParameters, WalletCli, WalletCommand,
+        restore_command, set_node_url_command, set_pow_command, sync_command, unlock_wallet, InitParameters, WalletCli,
+        WalletCommand,
     },
     error::Error,
     helper::{get_account_alias, get_decision, get_password, pick_account},
@@ -52,6 +53,13 @@ pub async fn new_wallet(cli: WalletCli) -> Result<(Option<Wallet>, Option<Accoun
             }
             WalletCommand::SetNodeUrl { url } => {
                 let wallet = set_node_url_command(storage_path, snapshot_path, url).await?;
+                (Some(wallet), None)
+            }
+            WalletCommand::SetPow {
+                local_pow,
+                worker_count,
+            } => {
+                let wallet = set_pow_command(storage_path, snapshot_path, local_pow, worker_count).await?;
                 (Some(wallet), None)
             }
             WalletCommand::Sync => {